# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Workflow Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SendOtpRequestPurpose =
  (typeof SendOtpRequestPurpose)[keyof typeof SendOtpRequestPurpose];

export const SendOtpRequestPurpose = {
  login: "login",
  register: "register",
} as const;

export interface SendOtpRequest {
  mobile: string;
  purpose: SendOtpRequestPurpose;
}

export interface SendOtpResponse {
  message: string;
  expiresInSeconds: number;
}

export interface VerifyOtpRequest {
  mobile: string;
  otp: string;
}

export type AuthResponseRedirectTo =
  (typeof AuthResponseRedirectTo)[keyof typeof AuthResponseRedirectTo];

export const AuthResponseRedirectTo = {
  doctor_dashboard: "doctor_dashboard",
  reception_dashboard: "reception_dashboard",
  admin_panel: "admin_panel",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  doctor: "doctor",
  receptionist: "receptionist",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  mobile: string;
  role: UserRole;
  /** @nullable */
  clinicId?: number | null;
  /** @nullable */
  clinicCode?: string | null;
}

export interface Clinic {
  id: number;
  clinicCode: string;
  clinicName: string;
  clinicAddress: string;
  doctorName: string;
  doctorQualification: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  logoUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
  clinic?: Clinic;
  redirectTo: AuthResponseRedirectTo;
}

export interface RegisterClinicRequest {
  clinicName: string;
  clinicAddress: string;
  doctorName: string;
  doctorQualification: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  templateId?: number | null;
  otp: string;
}

export interface RegisterClinicResponse {
  clinic: Clinic;
  user: User;
  token: string;
  clinicCode: string;
}

export interface UpdateTemplateRequest {
  templateId: number;
}

export interface PrescriptionTemplate {
  id: number;
  name: string;
  description: string;
  previewColor: string;
  headerStyle: string;
  footerStyle: string;
  bodyLayout: string;
  isDefault: boolean;
}