# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clinic Workflow Automation API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Send OTP to mobile number
 */
export const SendOtpBody = zod.object({
  mobile: zod.string(),
  purpose: zod.enum(["login", "register"]),
});

export const SendOtpResponse = zod.object({
  message: zod.string(),
  expiresInSeconds: zod.number(),
});

/**
 * @summary Verify OTP and login
 */
export const VerifyOtpBody = zod.object({
  mobile: zod.string(),
  otp: zod.string(),
});

export const VerifyOtpResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    mobile: zod.string(),
    role: zod.enum(["doctor", "receptionist", "admin"]),
    clinicId: zod.number().nullish(),
    clinicCode: zod.string().nullish(),
  }),
  clinic: zod
    .object({
      id: zod.number(),
      clinicCode: zod.string(),
      clinicName: zod.string(),
      clinicAddress: zod.string(),
      doctorName: zod.string(),
      doctorQualification: zod.string(),
      mobile: zod.string(),
      email: zod.string().nullish(),
      templateId: zod.number().nullish(),
      logoUrl: zod.string().nullish(),
      createdAt: zod.string(),
    })
    .optional(),
  redirectTo: zod.enum([
    "doctor_dashboard",
    "reception_dashboard",
    "admin_panel",
  ]),
});

/**
 * @summary Register a new clinic
 */
export const RegisterClinicBody = zod.object({
  clinicName: zod.string(),
  clinicAddress: zod.string(),
  doctorName: zod.string(),
  doctorQualification: zod.string(),
  mobile: zod.string(),
  email: zod.string().nullish(),
  templateId: zod.number().nullish(),
  otp: zod.string(),
});

/**
 * @summary Get clinic details
 */
export const GetClinicParams = zod.object({
  clinicId: zod.coerce.number(),
});

export const GetClinicResponse = zod.object({
  id: zod.number(),
  clinicCode: zod.string(),
  clinicName: zod.string(),
  clinicAddress: zod.string(),
  doctorName: zod.string(),
  doctorQualification: zod.string(),
  mobile: zod.string(),
  email: zod.string().nullish(),
  templateId: zod.number().nullish(),
  logoUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Update clinic prescription template
 */
export const UpdateClinicTemplateParams = zod.object({
  clinicId: zod.coerce.number(),
});

export const UpdateClinicTemplateBody = zod.object({
  templateId: zod.number(),
});

export const UpdateClinicTemplateResponse = zod.object({
  id: zod.number(),
  clinicCode: zod.string(),
  clinicName: zod.string(),
  clinicAddress: zod.string(),
  doctorName: zod.string(),
  doctorQualification: zod.string(),
  mobile: zod.string(),
  email: zod.string().nullish(),
  templateId: zod.number().nullish(),
  logoUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary List all available prescription templates
 */
export const ListTemplatesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  previewColor: zod.string(),
  headerStyle: zod.string(),
  footerStyle: zod.string(),
  bodyLayout: zod.string(),
  isDefault: zod.boolean(),
});
export const ListTemplatesResponse = zod.array(ListTemplatesResponseItem);